Use Coercion in the Class hierarchy.
This change was unavoidable because implicit arguments were cluttering
the proof context. One downside is that some proofs no longer proceed
using just auto. Instead we have to explicitly name the lemma that is
available via a class instance in the context, and use the 'c' tactic
to resolve the coercion.