I proved Star_Semiring mat.
I finally decided that proving the validity of mat_plus_WFK is too
difficult, at least for now. A spree of new lemmas would be needed
to deal with the sub-matrix selection, and induction on the number
of steps turned out to be difficult due to the entanglement of
dependent vector and number types. Apart from mat_plus_WFK having
a shorter definition, it is not clear if it is even more efficient
than mat_plus_ind.